moved relevant functions to bounding_box.py
